Fix size_t/int related warnings

+++ src/scxml/ScXMLEventTarget.cpp
@@ -651,7 +651,7 @@ ScXMLEventTarget::sendExternalEvent(const ScXMLSendElt * sendelt)
   const char * delaystr = sendelt->getDelayAttribute();
   if (delaystr) {
     // format: "<num>s" or "<num>ms"
-    int len = strlen(delaystr);
+    size_t len = strlen(delaystr);
     if (len > 1) {
       if (delaystr[len-1] != 's') {
         SoDebugError::post("ScXMLEventTarget::sendExternalEvent",
